+++ src/app/app.module.ts
@@ -2,10 +2,11 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { AboutComponent } from './components/about/about.component';
 import { CartModule } from './components/carts/cart.module';
 import { OrderModule } from './components/orders/order.module';
 import { ProductModule } from './components/products/product.module';
-import { AboutComponent } from './components/about/about.component';
+import { SharedModule } from './components/shared/shared.module';
 
 @NgModule({
     declarations: [
@@ -18,7 +19,8 @@ import { AboutComponent } from './components/about/about.component';
 
         ProductModule,
         OrderModule,
-        CartModule
+        CartModule,
+        SharedModule
     ],
     bootstrap: [AppComponent]
 })