+++ src/app/components/shared/shared.module.ts
@@ -1,12 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { DivFontSizeDirective } from './directives/div-font-size.directive';
 import { HighlightDirective } from './directives/highlight.directive';
 
 
 
 @NgModule({
-    declarations: [HighlightDirective],
+    declarations: [HighlightDirective, DivFontSizeDirective],
     imports: [CommonModule],
-    exports: [HighlightDirective]
+    exports: [HighlightDirective, DivFontSizeDirective]
 })
 export class SharedModule { }